requirements.txt
--extra-index-url https://git.fs.tum.de/api/v4/groups/django/-/packages/pypi/simple
Django~=4.0
django-active-link~=0.1.8
django-bootstrap-datepicker-plus~=4.0.0
django-bootstrap-modal-forms~=2.2.0
django-bootstrap5~=21.3
django-compref-keycloak~=1.2.0
django-crontab~=0.7.1
django-ical~=1.8.3
django-modeltranslation~=0.17.5
django-tex~= 1.1.10
icalendar~=4.0.9
Pillow~=9.1.0
python-dateutil~=2.8.2
qrcode~=7.3
